somewhat functional implementation of sprintf

+++ include/lolint.h
@@ -4,12 +4,12 @@ typedef __SIZE_TYPE__ size_t;
 
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
-typedef unsigned long uint32_t;
+typedef unsigned int uint32_t;
 typedef unsigned long long uint64_t;
 typedef __UINTPTR_TYPE__ uintptr_t;
 
 typedef char int8_t;
 typedef short int16_t;
-typedef long int32_t;
+typedef int int32_t;
 typedef long long int64_t;
 typedef __INTPTR_TYPE__ intptr_t;
